+++ Assets/Project/Programmer/MKH/Scripts/Test/BlueChipChoicePopUpTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a82a4bd079b17cd448870a414d51b2f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
